CV_DL/HW4-Image-Super-Resolution/requirements.txt
numpy>=1.18.5
pandas==1.0.1
torch>=1.6.0
torchvision>=0.7.0
opencv-python==4.2.0
more-itertools==8.4.0
Pillow==7.0.0
matplotlib>=3.2.2
glob3
tqdm>=4.41.0
pycocotools==2.0.2
scikit-learn==0.23.2
pretrainedmodels==0.7.4
jupyter==1.0.0
efficientnet-pytorch==0.7.0


